Reverted ForgotPasswordStep2 to it's initial condition. Created new Step for Code
